Add CLI option to disable OpenCL.
